Refactor Expr 1) moving Hamiltonian/StepHamiltonian and ThreadedMatrix to Operator module 2) moving all linalg associate to 1) are moving to Operator module.
