+++ lib/BloqadeExpr/src/Operator/preferences.jl
